LayoutTests/http/tests/security/xssAuditor/cached-frame-expected.txt

CONSOLE MESSAGE: line 5: Refused to execute a JavaScript script. Source code of script found within request.

CONSOLE MESSAGE: line 5: Refused to execute a JavaScript script. Source code of script found within request.

Check that an X-XSS-Protection header added by a 304 response does not override one from the original request.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Two console messages should be generated, noting that JavaScript was blocked.
Check that the nonce is the same, meaning that the document was only generated once:
PASS frame1.contentDocument.querySelector("input").value == frame2.contentDocument.querySelector("input").value is true
PASS successfullyParsed is true

TEST COMPLETE
  
